feat: promoting to mod and introducing stream capacity
