+++ crates/exex/exex/src/lib.rs
@@ -37,9 +37,6 @@
 mod backfill;
 pub use backfill::*;
 
-mod backfill_stream;
-pub use backfill_stream::*;
-
 mod context;
 pub use context::*;
